chore: rename group component files
