+++ client/src/app/_components/groups/GroupPostsList.tsx
@@ -4,8 +4,8 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import React, { useEffect, useState } from 'react';
 import { api } from '~/trpc/react';
-import { type FirebasePostModel } from '../../../../models/firebase-post-model';
-import GroupPostItem from '../group-post-item';
+import { type FirebasePostModel } from '../../../models/firebase-post-model';
+import GroupPostItem from './GroupPostItem';
 
 type GroupPostsListProps = {
 	groupId: string;
