+++ client/src/pages/groups/[groupId]/index.tsx
@@ -1,8 +1,8 @@
 import { useRouter } from 'next/router';
 import { useState } from 'react';
-import GroupPost from '~/app/_components/groups/group-post';
+import GroupPost from '~/app/_components/groups/GroupPost';
 // import GroupPostInput from '~/app/_components/groups/group-post-input';
-import GroupPostsList from '~/app/_components/groups/group-posts-list';
+import GroupPostsList from '~/app/_components/groups/GroupPostsList';
 import { BasicButton } from '~/app/_components/ui/basic-button';
 import { PageHeader } from '~/app/_components/ui/page-header';
 import PlatformLayout from '~/layouts/platform';
